requirements.txt
requests
python-dotenv
pydantic
pymongo
flask
flask_pymongo
Flask-JWT-Extended
flask-cors
flask[async]
pymongo[srv]